User

+++ app.py
@@ -1,18 +1,103 @@
+from datetime import datetime, timedelta, timezone
+from functools import wraps
+
+import jwt
 from flask import Flask, jsonify, request
+from werkzeug.security import check_password_hash, generate_password_hash
 
 from . import app
-from .models import Character, db
+from .models import Character, User, db
 
 
-# Rota para criar um novo personagem
+# Função de verificação de token
+def token_required(f):
+    @wraps(f)
+    def decorated(*args, **kwargs):
+        token = None
+        if "Authorization" in request.headers:
+            token = request.headers["Authorization"].split(" ")[1]  # O token está no formato "Bearer <token>"
+        if not token:
+            return jsonify({"message": "Token não encontrado."}), 401
+
+        try:
+            data = jwt.decode(token, "how-to-find-true-love-and-happiness", algorithms=["HS256"])
+            current_user = User.query.filter_by(id=data["id"]).first()
+        except Exception as e:
+            print(e)
+            return jsonify({"message": "Token inválido."}), 401
+
+        return f(current_user, *args, **kwargs)
+
+    return decorated
+
+
+# Rota para criar um novo usuário
+@app.route("/users", methods=["POST"])
+def signup():
+    data = request.get_json()
+    username = data.get("username")
+    password = data.get("password")
+
+    if username and password:
+        user = User.query.filter_by(username=username).first()
+        if user:
+            return (
+                jsonify({"message": "Usuário já existe. Por favor, faça login."}),
+                200,
+            )
+        user = User(username=username, password=generate_password_hash(password))
+        db.session.add(user)
+        db.session.commit()
+        return jsonify({"id": user.id, "username": user.username}), 201
+    return jsonify({"message": "Usuário ou senha inválidos."}), 400
+
+
+# Rota para fazer login
+@app.route("/login", methods=["POST"])
+def login():
+    auth = request.json
+
+    if not auth or not auth.get("username") or not auth.get("password"):
+        return (
+            jsonify({"message": "Usuário ou senha inválidos."}),
+            401,
+        )
+
+    user = User.query.filter_by(username=auth.get("username")).first()
+
+    if not user:
+        return (
+            jsonify({"message": "Usuário ou senha inválidos."}),
+            401,
+        )
+    if check_password_hash(user.password, auth.get("password")):
+        token = jwt.encode(
+            {
+                "id": user.id,
+                "exp": datetime.now(timezone.utc) + timedelta(minutes=30),
+            },
+            "how-to-find-true-love-and-happiness",
+            "HS256",
+        )
+        return jsonify({"token": token}, 201)
+    return jsonify({"message": "Usuário ou senha inválidos."}), 401
+
+
+# Rota para criar um novo personagem (agora vinculado ao usuário)
 @app.route("/characters", methods=["POST"])
-def create_character():
+@token_required
+def create_character(current_user):
     data = request.get_json()
     new_character = Character(
-        name=data["name"], race=data["race"], class_=data["class_"], level=data["level"]
+        name=data["name"],
+        race=data["race"],
+        class_=data["class_"],
+        level=data["level"],
+        user_id=current_user.id,
     )
     db.session.add(new_character)
     db.session.commit()
+
     return (
         jsonify(
             {
@@ -21,16 +106,22 @@ def create_character():
                 "race": new_character.race,
                 "class": new_character.class_,
                 "level": new_character.level,
+                "user_id": new_character.user_id,
             }
         ),
         201,
     )
 
 
-# Rota para listar todos os personagens
+# Rota para listar todos os personagens do usuário logado
 @app.route("/characters", methods=["GET"])
-def get_characters():
-    characters = Character.query.all()
+@token_required
+def get_characters(current_user):
+    characters = current_user.characters
+
+    if not characters:
+        return jsonify({"message": "Usuário não possui personagens."}), 404
+
     return (
         jsonify(
             [
@@ -48,10 +139,15 @@ def get_characters():
     )
 
 
-# Rota para buscar um personagem específico por ID
+# Rota para buscar um personagem específico do usuário logado por ID
 @app.route("/characters/<int:id>", methods=["GET"])
-def get_character(id):
-    character = Character.query.get_or_404(id)
+@token_required
+def get_character(current_user, id):
+    character = Character.query.filter_by(id=id, user_id=current_user.id).first()
+
+    if not character:
+        return jsonify({"message": "Personagem não encontrado."}), 404
+
     return (
         jsonify(
             {
@@ -68,8 +164,13 @@ def get_character(id):
 
 # Rota para editar um personagem existente
 @app.route("/characters/<int:id>", methods=["PUT"])
-def update_character(id):
-    character = Character.query.get_or_404(id)
+@token_required
+def update_character(current_user, id):
+    character = Character.query.filter_by(id=id, user_id=current_user.id).first()
+
+    if not character:
+        return jsonify({"message": "Personagem não encontrado."}), 404
+
     data = request.get_json()
 
     character.name = data.get("name", character.name)
@@ -92,10 +193,15 @@ def update_character(id):
     )
 
 
-# Rota para deletar um personagem
+# Rota para deletar um personagem do usuário
 @app.route("/characters/<int:id>", methods=["DELETE"])
-def delete_character(id):
-    character = Character.query.get_or_404(id)
+@token_required
+def delete_character(current_user, id):
+    character = Character.query.filter_by(id=id, user_id=current_user.id).first()
+
+    if not character:
+        return jsonify({"message": "Personagem não encontrado."}), 404
+
     db.session.delete(character)
     db.session.commit()
     return jsonify({"message": "Personagem deletado com sucesso."}), 200
